Fix warning on Linux

+++ src/window.cpp
@@ -7,9 +7,7 @@
 #include <fmt/core.h>
 
 Window::Window(int width, int height, const char* title)
-    : window_{glfwCreateWindow(static_cast<int>(width),
-                               static_cast<int>(height), title, nullptr,
-                               nullptr)}
+    : window_{glfwCreateWindow(width, height, title, nullptr, nullptr)}
 {
   if (!window_) {
     fmt::print(stderr, "Error: Failed to create a GLFW Window");
